requirements.txt
openai
azure-ai-evaluation
python-dotenv
bs4
promptflow[azure]
azure-identity